[IMP] knowledge: improve design
Purpose:
========
This commit improves the design and the wording of knowledge.

- Change the indentation of the articles names in the sidebar, so
that articles of the same level (parent/1st child/2nd child, ...)
are aligned. Now, carets are always displayed, but are not visible if
an article has no child.

- Make the size of the input match exactly the length of the article
name: previously, the input size was set using the size attribute,
which approximates the length using the number of characters of the
name and a mean character width value. This resulted in cropped names
if there was a majority of large characters, or in a too large input
if it was the opposite.
The input size now matches exactly the article name, at any time.

- Change the range of the allowed width for the resizable sidebar:
allows users to make it smaller than before, but not as wide. The
max width is now set to the "sm" breakpoint, so that every article
visible on large screen is visible on small screens as well. The min
width is set to the width of the emoji picker so that it is fully shown
when changing an emoji from the sidebar

- Change wording of messages shown when trying to move an article:
adds the article name in the message, and the destination in the
title to make it easier for the user

- Make icons bigger in the options panel

- Avoid changing the look of the sidebar when installing website:
fixes the font-size in the sidebar so that the search bar, section
headers and article names are not too large when installing website.

- Fix non-working scrollbar cursor in sidebar: reduce the size of the
grabbable sidebar resizer so that the vertical scrollbar can be used
with the cursor

- Truncate article name and breadcrumb items when they are too large.

- Move favorite button in the toolbar.

- Add "fa-user-plus" icon before the invite people input in the share
panel, so that we have a column of aligned icons.

- Remove "o_knowledge_emoji_picker_container" which is not used anymore
since the usage of the new emojiPicker component

Task-2858661

Part-of: odoo/enterprise#31852