[IMP] knowledge: improve sidebar resizer
Purpose:
========

- Add resizable sidebar in frontend: allows visitors to resize the
sidebar, the same way users can do it in backend, as visitors can also
have small screens

- Keep selected sidebar size: when an user changes the size of the
sidebar, the selected width is stored in localStorage so that it can be
set automatically when the user selects another article or reloads the
page. This has only been implemented in backend view, because it's not
possible to do it without having a flicker in frontend (because of  the
public widget lifecycle).

- Move common scss to `knowledge_common.scss` instead of having duplicates
in `knowledge_views.scss` and `knowledge_frontend.scss`

Task-2858661

X-original-commit: 9621d8a8e13f8800b449f0f409df7ebd8649ad07
Part-of: odoo/enterprise#32070