[FIX] account_reports: translate text on dialog box
Currently, if you change the language and then go to Accounting > Reporting > General Ledger and click on one entry > Annotate, the dialog box that appears is not translated.

This happens because _t() is not called on the appropriate strings.

This commit fixes the issue by adding _t() calls where needed.

opw-2831616

closes odoo/enterprise#29417

X-original-commit: 83133830fb7ac85d83ccba8719f22903f1eb18d8
Signed-off-by: Adrien Widart <awt@odoo.com>
Signed-off-by: Nshimiyimana Serge Séna (sesn) <sesn@odoo.com>