[PERF] knowledge: avoid computing ACLs of the active article twice
This commit improves the knowledge side panel performances by avoiding to
compute the active article ACLs before building the side-panel values.

They will be computed alongside all "visible articles" at once, benefiting from
batch computation of fields and removing extra queries.

The ancestors to unfold to reach the active article are now computed using a
sudo fetch of the active articles (to avoid ACLs queries) and built using the
parent_path field instead of actively fetching all ancestors.

The ancestors will not be displayed if not available to the end-user, as they
need to be part of all the "visible articles" (for which we fully check ACLs).

Task-2791000

Part-of: odoo/enterprise#29442