[IMP] knowledge: add a CoverSelector based on ImageSelector
Purpose:
========

Instead of using the standard "avatar" image picker for the covers,
create a CoverSelector, based on ImageSelector used in the MediaDialog
from web_editor which is prettier, allows to reuse covers, and allows to
select images from unsplash.

Specs:
======

Create a "knowledge.cover" model which references an image in ir.attachment
(the actual cover image).
This allows to create a "bank of images" for knowledge articles, so that
images uploaded as covers can be reused in different articles without
duplicating the attachment, but cannot be seen in the mediaDialog when
used in other apps, or when used in the body of articles.
The model also stores the url of the images, so that the images can be
displayed in frontend (these urls contains access tokens when they are
required for external access, which is the case for images uploaded using
the "upload image" button).

Create the components "CoverDialog" and "CoverSelector", based on
"MediaDialog" and "ImageSelector" but adapted for knowledge's use-cases
and the new "knowledge.cover" model.

The binary field "cover" is replaced by a many2one field "cover_image_id"
referencing a record of the new "knowledge.cover" model.
A new field "cover_image_id" has also been added, referencing the covers'
attachment, which is used to display the cover.

When a cover is not used anymore (not linked to any article), it is
deleted using an autovacuum.

Task-2852916

X-original-commit: a88bab9ac71945ae5bdb7f7576c8f898c1add4b6
Part-of: odoo/enterprise#32050