[IMP] knowledge: add random unpslash image on "add cover" click
Purpose:
========
When the user clicks on "add cover" in knowledge, picks a random
image from unsplash using the name of the article as search query.
If unsplash credentials are not set or if service is down, falls back to
the "change cover" behaviour, showing the CoverSelector allowing the
user to pick an existing cover, to upload a new one, or to enter its
unsplash credentials.

Specs:
======
- Extend "web_unsplash" controller and add a route that fetches a random
image from unsplash using a query word, then uses "web_unsplash"'s route
to save the image in ir_attachments.

- Add a method calling this route using the article name as query when
clicking on the "Add cover" button. During the request, the button is
hidden so that one can not click multiple times on it and send several
rpc's. If the request returns an error (unsplash credentials not set,
service unavailable, ...), opens the CoverSelectorDialog instead.

Task-2852916

Part-of: odoo/enterprise#31744