[IMP] spreadsheet: Apply currency formatting to lists
With the introduction of dynamic format for formulas, we can now apply
currency specific formats to list formulas (it required to fetch
res.currency data from the server) which will apply a dedicated format
for each user depedending on the fetched records.

Note that this feature does not apply to pivots as the aggregate can mix
up different currencies and we don't have a simple wsay to detect if a
single currency was used. For these cases, users can resort to a static
currency format.

Task 2788853

closes odoo/enterprise#29431

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Co-authored-by: Alexis Lacroix <laa@odoo.com>