[FIX] documents_project: fix traceback of a active model in documents
Before this commit, When we open the documents from the project.task
form view, and reloading the current page, there is a traceback
(active_model not defined) for active_model as there is no 'active_model'
in the url query string.

After this commit, we removed the active_model and replace it with static
model `project.task`.

task-2871552

closes odoo/enterprise#32502

X-original-commit: 9363ff61cb92616bbac925ea90ec1146dc1f0aa0
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>