[FIX] documents_project: fix traceback of a active model in documents
Before this commit, When we open the documents from the project.task
form view, and reloading the current page, there is a traceback
(active_model not defined) for active_model as there is no 'active_model'
in the url query string.

After this commit, we removed the active_model and replace it with static
model `project.task`.

task-2871552

closes odoo/enterprise#32575

X-original-commit: 1fb7ab4f8ded94cace6708eea6219cc63793a82a
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>