[FIX] documents_account: prevent tags from being overridden
Prior to this commit if a journal to synchronize was configured in the
document_account settings was setup it would always override the tags
regardless of prior state.
This means that for documents created by workflow actions the tags would
simply not work properly for example with the action "Create Bill".

OPW-2909133

closes odoo/enterprise#29521

X-original-commit: 70d4fa07f8863ee2f65315c125989f81acb5c21a
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>