[PERF] knowledge: batch compute user write access for the tree panel
This commit improves the performance of the side tree panel by pre-computing
the 'user_has_write_access' field for all articles before going into the
template rendering.

This allows to batch the computation before the rendering as opposed to
computing the field value for every article displayed in the tree
independently.

Task-2791000

Part-of: odoo/enterprise#29442