[IMP] knowledge: add trash management
=======
Purpose
=======

Allow users to get rid of articles they do not need anymore while adding a
safeguard against irreversible deletions.

============
Specifications
============

In the kebab menu, replace the Archive action by a "Send to Trash" button

This new action does the following:
-> It archives the articles
-> It gives them a "deletion date" of NOW+30days, this delay being configurable
   through ``knowledge_article_trash_limit_days`` config parameter

Add a discrete Trash button in the end of the side panel of home view.
=> It sends to a list view of trashed articles
(the filter can also be available by default in the search view)

When a trashed article is opened, add a ribbon at the top of the form view that
says so + allows to reverse this operation

Removing article from trash make it root article to ensure the article will be
displayed in the sidebar (in case its old parent is still archived).

To allow other users to react to this, we will notify them when an article they
follow is moved to the trash bin using a simple notification by email using
the new email template layout as for assignations, ...)

Note: Only one mail is send per user even if multiple articles are sent to
trashed.

Task-2850521 (Knowledge: trash management)

Part-of: odoo/enterprise#30908