[PERF] knowledge: avoid fetching favorited articles separately
This commit improves the performance of the side tree panel by fetching
favorited articles along with all the visible articles instead of fetching them
separately.

This allows fetching the knowledge.article.favorite linked to the current user
as sudo as we do not leak information, saving a few extra queries.

Task-2791000

Part-of: odoo/enterprise#29442