[FIX] hr_contract_salary: fix access to field_description
Since e7f94c7947698e0432dfef2020b3f6eb85168992, a resume explanation is
computed by concatenating the field description of the aggregated
advantages.
However, if the field description is not loaded in the cache yet, the
user could receive an AccessError when reading the field_description.

This commit adds a sudo() on this read so that the name is always read
correctly.

closes odoo/enterprise#29504

X-original-commit: ee281ab0124c9206ee10b83095f4bf2fda100782
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Paul Morelle (pmo) <pmo@odoo.com>