[IMP] account_sepa: add validation check on BIC code in SCT files
When generating SEPA Credit Transfer XML files, currently there is no
validation check on BIC codes. This leads to files generated being
rejected by banks afterwards as the BIC code must fulfill the XSD
requirements. The purpose of this commit is to add that requirement at
the generation of the file, so the customer can understand and fix it
himself.

closes odoo/enterprise#31328

X-original-commit: 32de1654c76c5640129ebef235eddaf3f0d8e1a1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>