[IMP] account_online_synchronization: improve bank addition flow
When synchronizing with a bank to fetch transactions in Odoo, we have a
wizard that pop up at the end of the linking process so that the user
can link the fetched accounts with a journal. Actually the user must
select what to do for each account (create a new journal or linked with
an existing journal). What we want to do is to ease the flow by using
the following rules.

- if a journal with a similar bank account exists, we assign it to
    the corresponding new online account.

- for the remaining online accounts, we assing empty journals with no
    bank account set.

- when syncing the online accounts (last step),
    online accounts with no journal set get newly created journals.

The fetched online account currency is used to link the account with
a more appropriate journal. If not already, the currency gets activated.

task-2877418

closes odoo/enterprise#29613

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>