[FIX] sale_renting: correct price_unit when editing line before saving
Steps to reproduce:
1) config
- a client with a certain fiscal position
- a tax with tax included - at a certain rate e.g. 21%
- another tax with the same rate, but tax excl., with field 'Base affected by previous taxes' ticked
- on the fiscal position set on the client, a tax mapping from that 1st tax to that 2nd one
- a rental product on which the 1st tax - the one tax incl. - is applied
2) flow
- create a rental order
- select the rental product (the wizard will open)
- add 1 unit
- edit the line - click on the calendar (the wizard will open)

Issue:
The unit_price is not tax included

Cause:
When the rental order is not saved, fetching the tax from the first initialised line is not possible.
Therefore `product_taxes_after_fp = product_taxes` and the unit_price will be equal to its lst_price.
https://github.com/odoo/enterprise/blob/0359eff3ffacc62a0de6e5bbb0c8b144b8d2e781/sale_renting/wizard/rental_configurator.py#L104-L118

Solution:
When editing the line, give the `default_tax_ids` to the `_openRentalConfigurator`. Consequently, `unit_price` will be computed with the correct fpos.

opw-2893564

closes odoo/enterprise#29745

X-original-commit: 7dfffa96cc92f232e420123a0d63e0deaed91d83
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>