[FIX] account_saft: Fix issue with different SAF-T report versions
SAF-T reports have 2 standards: v1 and v2. Some countries are based on v1 while others on v2. In v1, DebitAmount or CreditAmount are optional, in v2 it's mandatory. This means that for countries using v2, this mandatory field might not always appear.
This commit addresses this issue by always displaying DebitAmount or CreditAmount even when both credit and debit are 0.0. The choice between the 2 depends on the move type (in_invoice or out_refund = DebitAmount, out_invoice or in_refund = CreditAmount)

task id=2920040

closes odoo/enterprise#30996

X-original-commit: 510e7c80317486bfe8f14c980fd6a93603ea277e
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Popeler Antoine (anpp) <anpp@odoo.com>