[FIX] knowledge: prevent crash on window resize with open chatter
Fix the bug where an error message would pop up when resizing the window in the
article form view with the chatter open.

`this._chatterContainerTarget` was not set up in the knowledge form renderer and
is required by the mail form renderer. This commit properly set up the target
when the chatter is rendered.

Task-2917593

closes odoo/enterprise#29385

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>