[FIX] sale_subscription: correct rights verification for order token assignment
If you access (and pay) for a SO with the portal access token, but you
don't have
generic write rights on it, the token assignment step will raise.

This commit (and its community counterpart) make sure the assign_token
route
follows the generic behavior of portal routes, relying on
_document_check_access to make sure the user is allowed to access (&
edit)
the record.

closes odoo/enterprise#28595

Related: odoo/odoo#94020
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>