[FIX] Worksheet: QWeb report view not updated
When going to Quality / Configuration / Quality Worksheet template, we can create and design one with studio.
The problem is that if industry_fsm_report is not installed, the QWeb report view will not be updated.
industry_fsm_report is not and should not be required for worksheets to work properly.
The solution found is to add a controller to the worksheets, just like it was done on the industry_fsm_report.

Should be forwarded
OPW-2867753

closes odoo/enterprise#29940

X-original-commit: 44040ded6bb68dbc5c14cd65128fc5de762ffd3b
Signed-off-by: William Henrotin (whe) <whe@odoo.com>