[IMP] helpdesk_sale_timesheet: move the SLA selection under the sales tab
This commit just changes the location of the SLA Policy selection in the product form.
Previously the SLA Policy selection was under the "General Information" tab but
now the selection has been moved under the "Sales" tab.

The imporvements in this commit aim at making the placement of the SLA policy setting
more intuitive for the users.

task-2919785

closes odoo/enterprise#29732

Signed-off-by: Xavier <xbo@odoo.com>