[FIX] knowledge: reset permission for root moved under parent
Previously, when moving a root article as a child of another article in
workspace, the internal permission was not reset, meaning that a private root
would keep `internal_permission: 'none'` when moved under a workspace article.

This fix forces the internal_permission reset when moving a root under a
'shared' or 'workspace' parent.

Task-28670529

X-original-commit: b78448b81d2856ccef8b51413d0e6490d85ba414
Part-of: odoo/enterprise#30312