[FIX] website_sale_renting_product_configurator: fix context keys
Due to some changes the route was not functioning properly anymore, as
the 'context' arguments was directly parsed as the environment's
context.
This meant that the overrides that are done for `website_sale_renting`
weren't working anymore making it impossible to add rental products with
product configurator to the cart.

closes odoo/enterprise#29558

X-original-commit: 1979507ad9655b169f7c609ffc8cb48000e016c0
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>