[IMP] project_timesheet_forecast_sale: display 'N/A' on non-plannable cells
In Project Update's description, display 'N/A' in non-plannable cells.
The column 'Planned' contains, for each sale order line,
the sum of hours of planning slot related to it.
When the sold product is measured in unit, this is not possible.

task-2909337

closes odoo/enterprise#29565

Related: https://github.com/odoo/odoo/pull/96104
Related: odoo/odoo#96104
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>