[FIX] timesheet_grid: properly synchronize input value with state
Previously, owl would not overwrite the value of an input when patched
if the value computed in the render was the same as the previous render,
even if the value in the DOM was changed (by the user or
programmatically). This was fixed, but this can cause issues in
components that relied on the old behaviour such as the timer header
component. This commit simply puts the input value in a state and
updates it as necessary, so that it is not lost on render.

closes odoo/enterprise#29642

X-original-commit: 37020e429b791c7f90d0f0a535ae93bbf312b565
Signed-off-by: Géry Debongnie <ged@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>