[FIX] timesheet_grid : search in timesheets fails task w/o project
Steps to reproduce:
- Create a task not linked to a project
- Go to the timesheet grid and make a custom filter that either "contains"
a part of the task name either "does not contain" something random that
is not a part of the task name

Current behavior:
Error with traceback

Expected behavior:
The created task should not appear as only tasks linked to projects can
have timesheets

Explanation:
In the search function for timesheets we create a dictionnary of domains
for the tasks/projects called "domain_project_task", this dictionnary for
its tasks domain does not contain a condition to avoid tasks without
project (that should not be an option in a timesheet context). Later if
a task with no project is in the possibilities it can create an error in
the read_row_fake_value as task.project_id.name_get() returns an empty
list. To solve the issue we add an element to the domain to restrict the
search to task with project_id.

opw-2906233

closes odoo/enterprise#30180

X-original-commit: 6aad6a78a71868d7125283e3a227d0387f68e506
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>