[FIX] l10n_cl_edi_boletas: disable the cron tasks related to sending daily sales report and ask for its status
After this commit:

The cron tasks that send l10n-cl daily sales book reports, and ask for their status have an auto-disable
code from August 2nd, 2022. Since the requirement to send the daily sales book
will be derogated by the SII from August 1st 2022.
Thus, in versions 14.0 and 15.0 even when the daily sales book code is still
available, the cron task that sends this report will be automatically disabled
by this PR from august 2st 2022.

The cron task that checks for the status will be disabled as well but from August 2nd.
in order to ensure the status update for the July 2022 reports.

If for any reason the user still has pending reports (only reports for July 2022), he has the chance to enable this by hand.
asking status of l10n-cl daily sales book.

closes odoo/enterprise#29895

X-original-commit: aeffb05cb9b693b949cad798991e346598e0953d
Signed-off-by: Josse Colpaert <jco@odoo.com>