[FIX] voip: improve voip service and adapt PhoneField to OWL
The voip service is now thought of as the center point of voip to
handle the VOIP logic. It can tell if one can call, and one can make
a phonecall with the `call` method.

The phone_field has been adapted consequently and uses that service
to handle calling a number with VOIP.

Part-of: odoo/enterprise#23069