[IMP] stock_barcode: confirmation when add product
While processing an operation, if the user scans a product's barcode and
this product isn't expected, a new line will be created.
That means every time the user will scan the wrong barcode, it will
create an unwanted line, which can be a real pain to deal with.

To avoid that, a confirmation dialog will ask the user if they really
want to add the scanned product in the operation instead of directly
create a new line.

About the test tours, we scan unreserved products a lot of time.
Depending if the operation is supposed to work with reserved quantity or
not, some of tours' pickings are now immediate transfers, and for the
other, each time an unreserved product is scanned, the prompt will be
confirmed (thanks to a new helper, `confirmAddingUnreservedProduct`).

I took the opportunity to rename `test_receipt_from_scratch_with_lots_5`
into `test_receipt_with_sn_1` as this test works with a planned
transfer (so, not from scratch) and works with serial, not lots.

task-2666478

Part-of: odoo/enterprise#27547