[FIX] sale_planning: fix the default SOL on plannning
Before this commit, when we try to create a shift from
SO form view > hours to plan stat button, the default SOL of
type service is not set on shift.

So in this commit, pass the id of SOL instead of object
as default value of `default_sale_line_id`.

task-2873238

closes odoo/enterprise#29757

X-original-commit: 5332bed239da7a28b547d188155afbea1b9507ac
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Mahendrasinh Barad (mba) <mba@odoo.com>