[FIX] web_studio: font style don't apply correctly
Since BS5 migration, font in studio was not the same. This is because
it uses the SCSS variable `$font-family-base`, but in BS5 this variable
is set with:
```scss
$font-family-base: var(--#{$variable-prefix}font-sans-serif) !default;
```
but since it use `--#{$variable-prefix}` and there is no
`bootstrap_overridden.scss` for studio there was a missmatch and so no
font was set and use the default font set in the browser.

To avoid this, we set the `--#{$variable-prefix}` to empty to avoid
missmatch.

closes odoo/enterprise#29679

Signed-off-by: Pierre Paridans (app) <app@odoo.com>