[IMP] account_accountant: improve invoice UX
- Some changes in the Journal Items UX were made to have specific
views opened for specific types of records. This commit changes
the "View" button in the Journal Entry popover to have the same
behaviour.

- The "Statement(s)" button now opens the new reconciliation widget.

- The core open_move method is also moved from the account move line
model to the account move's one for more flexibility.

- When clicking on a statement Journal Entry from the Journal Items
list view, the new reconciliation widget is now correctly focused on the
selected line.

task-2893208

closes odoo/enterprise#29092

Related: odoo/odoo#94923
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>