[FIX] knowledge: handle readonly articles in the article tree
The coloration (on hover, on click) was not properly set up for readonly
articles.

Also, the toleranceElement for each nestedSortable (section of the article tree)
was too restrictive (only on .o_article_handle, which is a class only applied on
articles with write access for the current user). Revert back to select the
direct div descendants (of a li tag in the tree).

The toleranceElement is mandatory for the nestedSortable to function, and would
not render properly with the restrictive selector.

Task-2925588

X-original-commit: 0ffee1a0f523532a04b81f6f01d81b1f48d46743
Part-of: odoo/enterprise#29739