[IMP] spreadsheet: better date filter defaults
Since task 2800957, the defaults of the year date filter do not make sense
anymore (i.e. selecting "2022" will always filter on year 2022,
not the current year).

The purpose of this task is to replace the date filter default value
selection by a checkbox to automatically set the default value to the
current period (current year/quarter/month depending on the rangeType).

Odoo task 2900848

closes odoo/enterprise#29365

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>