[IMP] l10n_co_edi: added vendor document feature
In this commit, we have added the edi feature for `Vendor Bill` and `Vendor Credit Note`.

When the vendor can not send the invoice to the government,
we might need to do it for him.

For that, you can create a separate purchase journal where you fill in the
DIAN information. That way, the vendor bills on that journal be sent
to the government.

TaskID - 2522967

closes odoo/enterprise#30014

X-original-commit: b39f73e9887d9b1c0043e725a7445b64be35b523
Signed-off-by: Josse Colpaert <jco@odoo.com>