[FIX] l10n_lu_reports: fix balance sheet
The special date changer on LU_BS_319 and LU_BS_ABR319
are incorrect.
It says "At the beginning of the period".
For example, user wants to extract a mid-term Balance Sheet and P&L.
Fiscal year : 01/01/xx up to 12/31/xx
Balance Sheet as of 06/30/xx
At the beginning means : 06/01/xx
While he expects "from the beginning" : 01/01/xx

It is not enough to change the special date changer to
"from the beginning" or the items on P&L accounts will
be double counted in LU_BS_319 and LU_BS_321, respectively
"Profit and Loss brought forward" and "Profit and Loss for
the financial year".

Adapt the formula of LU_BS_319 and LU_BSABR_319, and
test_balance_sheet accordingly.

opw-2920822

closes odoo/enterprise#29741

X-original-commit: fb03fa3baa72fbf8f3baef7d6a113c54229ebff5
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>