[IMP] helpdesk_timesheet: fix timesheet description required in mobile
Currently, when user enter the timsheet in editable tree from the
task form view the description is not required and if keep empty then
it will add the '/' as default description but in the mobile form view
of the timesheet description is still required.

so in this commit, make description non required in timesheet form
view of mobile.

task-2918757

closes odoo/enterprise#36231

X-original-commit: 1f4e1495ae0c928340ed46d4eb06f415e2e27a8e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>