[REF] industry_fsm{_sale}: refactor fsm material product kanban in OWL
Since the new kanban view in OWL is recently available, the commit
converts the custom kanban view of `product.product` in OWL. The legacy
code is also removed since the new kanban view in OWL will replace it.
The widget "fsm_product_quantity" widget is also converted in OWL since
it is only used in the kanban view of `product.product`.

task-2917530

Part-of: odoo/enterprise#30173
Co-authored-by: Laurent Stukkens (LTU) <ltu@odoo.com>