[IMP] (l10n_be_)hr_contract_salary, hr(_contract_salary)_payroll: UX improvments
Purpose: 'hr.contract.salary.advantage' has the field 'name' that we also
show on salary configurator. Sometimes, we don't want to show 'name' for
some advantages on the configurator, and to do this, prior we left the
field empty. Which is not convenient when for example we check the list
of advantages.
To solve this issue, we add 'show_name' field in this commit. That allows
to populate the field, while keeping salary configurator they way we want.

Beyond that, the commit includes UX changes.

task - 2902100

closes odoo/enterprise#30513

Related: odoo/upgrade#3783
Signed-off-by: Kevin Baptiste <kba@odoo.com>