[IMP] documents_spreadsheet: add XLSX import
Add the possibility to convert .xlsx files into o_spreadsheet.

You can create a o_spreadsheet from a xlsx by trying to open a xlsx in
odoo from either the kanban view of documents, or from the document inspector.

Added code to documents.py to extract a xlsx after doing some basic checks
to verify that the file is indeed a xlsx. The extracted files are then dumped
in the DB as a json, and o_spreadsheet.js can convert these files into
an odoo spreadsheet.

Odoo task 2345236

closes odoo/enterprise#24727

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>