[IMP] helpdesk: convert kanban view to owl
This commit converts the helpdesk team dashboard and kanban view to owl.

During the conversion some python methods lose their `@api.model`
because when the user will clicks on a action available on the helpdesk
dashbaord, the `doActionButton` function in action service will now be
called to correctly markup the `help` content defined in the action.
Since the `doActionButton` is used instead of using the `call` method to
fetch the action and then calling `doAction`, the first argument
passing to the call_button route will always be the resIds. However, if
the method called by the call button route contained the `@api.model`
decorator, the first argument passed will not be removed and so the
method will get too much arguments. To avoid that and keeping the using
of `doActionButton`, the `@api.model` decorator defined on the methods
only called in the helpdesk dashboard is removed.

Task-2924239

closes odoo/enterprise#31685

Signed-off-by: Xavier <xbo@odoo.com>