[FIX] iot: fix iot longpolling service
The iot_longpolling service is extending the bus service while it does
not make any use of it. Before the introducton of the multi tab service,
the crosstab bus was calling stopPolling during the heartbeat. Now that
the logic for multiple tab has been moved, the crosstab bus subscribes to
the 'no_longer_main_tab' event. Before the multi tab, the stopPolling method
was never called from the crosstab bus since the startPolling method was
never called either. Now that the crosstab bus subscribes to those events
in its init method, the stopPolling method of the iot longpolling service
can be called by the crosstab bus, without the required argument (iot_ip).
This leads to errors. In order to solve this issue, let's remove the unneccessary
extension of the bus service.

closes odoo/enterprise#29748

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>