[IMP] sale_amazon: call order-related operations with an RDT
Before this commit, to get some restricted information, we needed to
call 2 (or 3 to get the gift information) additional operations using a
restricted data token.

This information is now available in the basic operations as long as we
get a RDT adapted for it. When querying Amazon, we now specify the
restricted elements that need to be added in the response of the basic
operations.

task-2927248

closes odoo/enterprise#30175

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>