[FIX] sign: right precision for float and monetary display
Steps to reproduce:
- Add in an auto fill partner field of type float, or monetary
 in the sign app and pull that onto a document template.

Current behavior:
The auto-fill does not take into consideration the field precision

Expected behvaior:
The auto-fill takes into consideration the field precision

Solution:
We use the convert_to_column function to get the right representation
even if it was not meant for this use it returns the string of a float
or a monetary field with re right precision. To our knoweledge this is
the only function that does that. Depending of the field type it will
get the digits precision for the float and the rounding precision for
the monetary field to give the right format.
opw-2920379

closes odoo/enterprise#30693

X-original-commit: 0f42d1df59599e47df36a628ce619c648d2f1b74
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>