[FIX] delivery_dhl : correct declared value when using international shipping
When using international shipping, value for product in declaration is taken in the currency which was specified to the product,
but the currency of declaration is in the currency of the sale order. So, if product is in one currency, for example euros, while
sales order is in another currency(dollars), in declaration the total_amount will be sale value in euros, but the name of currency will be dollars, so Odoo will send incorrect value of product to connector.

With this fix, we will check if the sale order exists and take it from sale order if there is one, since declared value should correspond to the value in rate_shipment_value function.
If there is no sale order, we will simply take it from product value and send this data to connector.

Merge of original pr-s #29453 and #29838

fixes opw-2826030

closes odoo/enterprise#29870

X-original-commit: eeb37b1d53cc2448fbcb78d6acdb2c8359521fd8
Signed-off-by: Tiffany Chang <tic@odoo.com>