[FIX] pos_iot: restricting key to confirm or cancel popup
The "pressing a specific key to confirm/cancel a popup" behavior was added but the problem was that
some popups only had one button which was either confirm or cancel. It was thus possible to close
those popups by bypassing the default behavior (e.g. possible to "cancel" a popup which only
has one button linked to an overridden confirm method and vice-versa).

closes odoo/enterprise#29957

X-original-commit: 6aac89db8617d15f4fe0f2f3a2c123533d24a663
Related: odoo/odoo#97069
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>