[IMP] account_online_synch: supports currency_code from proxy
The odoofin proxy now returns the currency code of the accounts along
other information. As we are using the data from the proxy to create
the online sync, and that we do not have a currency_code field for the
moment, it will be ignored until master.
In master, a change in the wizard will require the currency code which
will then have its own field.

Task id #2930451

closes odoo/enterprise#29890

X-original-commit: 5342ed07c60eaf739b6249cdafcdad4ddccf8c40
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>