[FIX] stock_barcode*: properly terminate tours
The barcode interface tours tend to end while in the barcode
form (with the numpad).

Add dedicated helpers for the cases of discarding & scan-validating
the form and use those where necessary (or useful, for the latter).

A special note needs to be made for the tour
`test_receipt_reserved_lots_multiloc_1` (test of the same name):

- opening the form view is not necessary as the scan already triggers
  a save
- however we need to wait for the scan and save to complete before
  ending the tour
- and because the form was opened without waiting for the scan, the
  tour had a race condition where it triggered the save twice,
  concurrently
- this led to the second move line being duplicated, which is why
  the Python test had 4 lines and a qty of 6 even though the reserved
  qty is only 4, hence having had to fix the post-tour testing

closes odoo/enterprise#29824

Related: odoo/documentation#2550
Related: odoo/odoo#96517
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>