[FIX] documents_project: fix the display of child workspace in side panel
Before this commit, When we set a workspace template that has two
children workspaces > create a project > on the documents left side panel,
the children workspaces should be displayed as children of the project's
workspace instead of at same level.

So in this commit, check condition properly with id instead of tumple of name
and id to display the child workspace at proper level.

task-2926348

closes odoo/enterprise#32751

X-original-commit: 50e0778b579b9790d720cbaa9f3ca298a85b7ecb
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>