[FIX] helpdesk: on hover pencil doesn't show up
In Helpdesk's Overview, on hover, the Daily Target should display a
pencil icon to highlight the editable aspect of the field.

This issue is due to the fact that `opacity-*` classes were manually
generated before and are now handled by BS5... which adds an
`!important` to them.

This commit has to also apply the `!important` to the :hover rule that
displays the icon.

closes odoo/enterprise#29910

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>