[FIX] account_avatax_sale: use Avatax amounts in tax_totals_json
The standard values are computed via `compute_all` on
`account.tax`. This will lead to the wrong result if a line on an SO
is (partially) exempt or not taxable.

Example of a tax returned by Avatax for a line which is not taxable:

{
  'description': '[S-PACK-S] Success Pack: Standard (50h) (End-User)',
  'details': [{
    'country': 'US',
    'exemptAmount': 0.0,
    'id': 0,
    'jurisCode': '36',
    'jurisName': 'NEW YORK',
    'jurisType': 'STA',
    'jurisdictionType': 'State',
    'liabilityType': 'Seller',
    'nonTaxableAmount': 100.0,
    'rate': 0.04,
    'rateType': 'General',
    'rateTypeCode': 'G',
    'region': 'NY',
    'reportingExemptUnits': 0.0,
    'reportingNonTaxableUnits': 100.0,
    'reportingTax': 0.0,
    'reportingTaxCalculated': 0.0,
    'reportingTaxableUnits': 0.0,
    'stateAssignedNo': '',
    'tax': 0.0,
    'taxCalculated': 0.0,
    'taxName': 'NY STATE TAX',
    'taxSubTypeId': 'S',
    'taxType': 'Sales',
    'taxableAmount': 0.0,
    ...
  }]
}

In this case NY STATE TAX is returned but the product in question is
not considered taxable ('tax' is 0.0, 'reportingNonTaxableUnits' is
the full amount). Because the returned rate is still 0.04 (the normal
rate for this tax), an account.tax record is created with an amount of
4% which results in Odoo calculating the wrong tax.

It's not possible to solve the problem by omitting these taxes in Odoo
because partial exemptions are possible (i.e. the 'rate' field is only
informational and we should never rely on it).

opw-2941694
opw-2928081
opw-2946895

closes odoo/enterprise#30884

X-original-commit: 96fc21cea37e8a3ea4a5ff816f691a8ca2e169d0
Signed-off-by: William André (wan) <wan@odoo.com>