[FIX] account_asset: fix wrong account from model
What are the steps to reproduce the issue?
1. Create an asset model with different Fixed Asset Account and Depreciation Account
2. Create an asset using the newly created asset model

What is the current behavior that you observe?
Fixed Asset Account and Depreciation Account are the same

What would be your expected behavior in this case?
The accounts should match the ones in the asset model

The solution:
Add the changes the fixed asset account in the onchange of the model.
If the fixed assets account is not defined in the model, it will put it empty in the asset,
which will trigger the compute to take the one from the depreciation account.
We will not keep the one that was defined on the asset, as the model should take precedence.
Also ensures that if the asset is generated from a bill,
we don't override it with the account from the model.

opw-2929933
opw-3203232

closes odoo/enterprise#48363

X-original-commit: 2858ba1cb560978a204d125ba916c3034375062a
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>