[FIX] account_avatax_sale: recompute when changing optional products
It's possible for a customer to add, delete and modify the quantity of
optional products via /my/quote. This fix makes sure tax is recomputed
via avatax whenever that happens.

It's not possible to test this @http.route function with MockRequest
like in account_avatax_sale_subscription because this module doesn't
depend on website. Instead this fix was tested by adding a tour that
adds, increments and deletes an optional product. The test then checks
if button_update_avatax was called the right amount of times.

opw-2946916

closes odoo/enterprise#30265

X-original-commit: fb97a6a3ca1c4839b62b8216da22f0579549d9e9
Signed-off-by: William André (wan) <wan@odoo.com>