[IMP] project_enteprise: add an automatic scheduling for tasks
Before this commit, when the user wants to plan some tasks in project
or Field Service app  in the tasks gantt view, he can use the
magnifying glass.
The problem is when he selects his tasks to plan and validate all selected
tasks will be planned to the day/hour (according to the range) in which he
selects the magnifying glass. Moreover, the tasks planned could be in
conflict with another tasks in a basic project if a user assigned into
those tasks has another task planned in the same for another project.

This commit gives the opportunity to user to have an automatic scheduling
for tasks in the gantt view of a project. The tasks will be scheduled
in accordance with the global leaves of the company and the hours
planned for the tasks. If the tasks has no planned hours, they are
scheduled by default for a full day on the day the user selected. If the
tasks are assigned to a specific user, the scheduling will take into
account the personnal leaves and the tasks already assigned
(even in another project) to the user. The planned dates also depend
of the tasks dependencies and the priority level of tasks.
Moreover the tasks which are labelled as important will be scheduled before
the others ones.

task-2605513

Part-of: odoo/enterprise#25250
Co-authored-by: Xavier BOL (xbo) <xbo@odoo.com>